// ui/src/generated/api/model/dtoProblemGrantResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoProblemGrantResponseRole } from "./dtoProblemGrantResponseRole";

export interface DtoProblemGrantResponse {
  groupId?: string;
  groupName?: string;
  id: number;
  role: DtoProblemGrantResponseRole;
  userId?: string;
  username?: string;
}