// ui/src/generated/api/model/dtoContestGrantRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoContestGrantRequestRole } from "./dtoContestGrantRequestRole.ts";

export interface DtoContestGrantRequest {
  group?: string;
  role: DtoContestGrantRequestRole;
  username?: string;
}