// ui/src/generated/api/model/dtoProblemGrantRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoProblemGrantRequestRole } from "./dtoProblemGrantRequestRole";

export interface DtoProblemGrantRequest {
  group?: string;
  role: DtoProblemGrantRequestRole;
  username?: string;
}