// ui/src/generated/api/model/dtoSetAccessRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoSetAccessRequestRole } from "./dtoSetAccessRequestRole";

export interface DtoSetAccessRequest {
  group?: string;
  role: DtoSetAccessRequestRole;
  username?: string;
}